Update `enb-bemxjst` to `2.x`
