+++ package.json
@@ -31,17 +31,17 @@
   },
   "peerDependencies": {
     "enb": ">= 0.11.0 < 1.0.0",
-    "enb-bem-i18n": "v1.0.0-beta2"
+    "enb-bem-i18n": "1.0.0-beta4"
   },
   "dependencies": {
-    "bemhtml-compat": "0.1.2",
-    "enb-bemxjst": "1.3.4",
+    "enb-bemxjst": "2.0.0-beta4",
     "vow": "0.4.9",
     "serialize-javascript": "1.0.0"
   },
   "devDependencies": {
     "bower": "1.3.9",
     "enb": ">= 0.11.0 < 1.0.0",
+    "enb-bem-i18n": "1.0.0-beta4",
     "istanbul": "0.3.14",
     "jscs": "1.13.1",
     "jshint": "2.7.0",
