+++ techs/bemhtml-i18n.js
@@ -1,28 +1,26 @@
 var EOL = require('os').EOL,
-    path = require('path'),
     vow = require('vow'),
-    vfs = require('enb/lib/fs/async-fs'),
-    bemcompat = require('bemhtml-compat'),
-    asyncRequire = require('enb/lib/fs/async-require'),
-    dropRequireCache = require('enb/lib/fs/drop-require-cache'),
     keysets = require('enb-bem-i18n/lib/keysets'),
-    compile = require('enb-bem-i18n/lib/compile'),
-    XJST_SUFFIX = 'xjst';
+    compileI18N = require('enb-bem-i18n/lib/compile');
 
 /**
  * @class BemhtmlI18nTech
  * @augments {BemhtmlTech}
  * @classdesc
  *
- * Compiles localized BEMHTML template files with BEMXJST translator and merges them into a single BEMHTML bundle.<br/>
- * <br/>
- * Localization is based on pre-built `?.keysets.{lang}.js` bundle files.<br/><br/>
+ * Compiles localized BEMHTML template files with BEMXJST translator and merges them into a single BEMHTML bundle.
  *
- * Important: It supports only JavaScript syntax by default. Use `compat` option to support old BEMHTML syntax.
+ * Localization is based on pre-built `?.keysets.{lang}.js` bundle files.
  *
- * @param {Object}    [options]                                    Options
+ * Important: It supports only JS syntax by default. Use `compat` option to support old BEMHTML syntax.
+ *
+ * @param {Object}    options                                      Options
  * @param {String}    [options.target='?.bemhtml.{lang}.js']       Path to a target with compiled file.
+ * @param {String}    [options.filesTarget='?.files']              Path to a target with BEMHTML FileList.
+ * @param {String[]}  [options.sourceSuffixes]                     Files with specified BEMHTML suffixes
+ *                                                                 involved in the assembly.
  * @param {String}    options.lang                                 Language identifier.
+ * @param {String}    [options.keysetsFile='?.keysets.{lang}.js']  Path to a source keysets file.
  * @param {String}    [options.exportName='BEMHTML']               Name of BEMHTML template variable.
  * @param {Boolean}   [options.compat=false]                       Sets `compat` option to support old BEMHTML syntax.
  * @param {Boolean}   [options.devMode=true]                       Sets `devMode` option for convenient debugging.
@@ -32,16 +30,16 @@ var EOL = require('os').EOL,
  * @param {Boolean}   [options.cache=false]                        Sets `cache` option for cache usage.
  * @param {Object}    [options.requires]                           Names of dependencies which should be available from
  *                                                                 code of templates.
- * @param {String[]}  [options.sourceSuffixes]                     Files with specified suffixes involved in the
- *                                                                 assembly.
- * @param {String}    [options.keysetsFile='?.keysets.{lang}.js']  Path to a source keysets file.
  *
  * @example
- * var BemhtmlI18nTech = require('enb-bem-i18n/techs/bemxjst/bemhtml-i18n'),
+ * var BemhtmlI18nTech = require('enb-bemxjst-i18n/techs/bemhtml-i18n'),
+ *     KeysetsTech = require('enb-bem-i18n/techs/keysets'),
  *     FileProvideTech = require('enb/techs/file-provider'),
  *     bem = require('enb-bem-techs');
  *
  * module.exports = function(config) {
+ *     config.setLanguages(['en', 'ru']);
+ *
  *     config.node('bundle', function(node) {
  *         // get FileList
  *         node.addTechs([
@@ -52,12 +50,10 @@ var EOL = require('os').EOL,
  *         ]);
  *
  *         // collect and merge keysets files into bundle
- *         node.addTechs([
- *            [ Keysets, { lang: '{lang}' } ]
- *         ]);
+ *         node.addTech([KeysetsTech, { lang: '{lang}' }]);
  *
- *         // build localized BEMHTML file for given {lang}
- *         node.addTech([ BemhtmlI18nTech, { lang: '{lang}' } ]);
+ *         // build localized BEMHTML file for each lang
+ *         node.addTech([BemhtmlI18nTech, { lang: '{lang}' }]);
  *         node.addTarget('?.bemhtml.{lang}.js');
  *     });
  * };
@@ -66,55 +62,78 @@ module.exports = require('enb-bemxjst/techs/bemhtml').buildFlow()
     .name('bemhtml-i18n')
     .target('target', '?.bemhtml.{lang}.js')
     .defineRequiredOption('lang')
-    .useFileList(['bemhtml', 'bemhtml.xjst'])
     .useSourceFilename('keysetsFile', '?.keysets.{lang}.js')
-    .builder(function (sourceFiles, keysetsFilename) {
-        var cache = this.node.getNodeCache(this._target),
-            cacheKey = 'keysets-file-' + path.basename(keysetsFilename),
-            promise;
-
-        if (cache.needRebuildFile(cacheKey, keysetsFilename)) {
-            dropRequireCache(require, keysetsFilename);
-            promise = asyncRequire(keysetsFilename)
-                .then(function (keysets) {
-                    cache.cacheFileInfo(cacheKey, keysetsFilename);
-
-                    return keysets;
-                });
-        } else {
-            promise = asyncRequire(keysetsFilename);
+    .builder(function (fileList, keysetsFilename) {
+        // don't add fat wrapper code of bem-xjst
+        if (fileList.length === 0) {
+            return this._mockBEMHTML();
         }
 
-        return promise
-            .then(function (sources) {
-                var parsed = keysets.parse(sources),
-                    template = [
-                        'oninit(function(exports, context) {',
-                        '    context.BEMContext.prototype.i18n = ' + compile(parsed, this._lang) + ';',
-                        '});'
-                    ].join(EOL);
+        return vow.all([
+            this._getBEMHTMLSources(fileList),
+            this._compileI18N(keysetsFilename)
+        ]).spread(function (BEMHTMLSources, I18NCode) {
+            // i18n will be available in templates by `this.i18n`
+            var sources = BEMHTMLSources.concat({
+                contents: [
+                    'oninit(function(exports, context) {',
+                    '    var BEMContext = exports.BEMContext || context.BEMContext;',
+                    '    BEMContext.prototype.i18n = ' + I18NCode + ';',
+                    '});'
+                ].join(EOL)
+            });
 
-                return this._readSourceFiles(sourceFiles, true)
-                    .then(function (sources) {
-                        var source = sources.join(EOL) + template;
-                        return this._bemxjstProcess(source);
-                    }, this);
-            }, this);
+            return this._compileBEMXJST(sources);
+        }, this);
     })
     .methods({
-        _readSourceFiles: function (sourceFiles, oldSyntax) {
-            return vow.all(sourceFiles.map(function (file) {
-                return vfs.read(file.fullname, 'utf8')
-                    .then(function (source) {
-                        if (oldSyntax && XJST_SUFFIX !== file.suffix.split('.').pop()) {
-                            source = bemcompat.transpile(source);
-                        }
+        /**
+         * Reads source code of BEMHTML templates and processes.
+         *
+         * @param {FileList} fileList — objects that contain file information.
+         * @returns {Promise}
+         * @private
+         */
+        _getBEMHTMLSources: function (fileList) {
+            var filenames = this._getUniqueFilenames(fileList);
 
-                        return '/* begin: ' + file.fullname + ' */' + EOL +
-                            source +
-                            EOL + '/* end: ' + file.fullname + ' *' + '/';
-                    });
-            }));
+            return this._readFiles(filenames)
+                .then(this._processSources, this);
+        },
+        /**
+         * Compiles i18n module.
+         *
+         * Wraps compiled code for usage with different modular systems.
+         *
+         * @param {String} keysetsFilename — path to file with keysets..
+         * @returns {Promise}
+         * @private
+         */
+        _compileI18N: function (keysetsFilename) {
+            return this._readKeysetsFile(keysetsFilename)
+                .then(function (keysetsSource) {
+                    var parsed = keysets.parse(keysetsSource),
+                        opts = {
+                            version: parsed.version,
+                            language: this._lang
+                        };
+
+                    return compileI18N(parsed.core, parsed.keysets, opts);
+                });
+        },
+        /**
+         * Reads file with keysets.
+         *
+         * @param {String} filename — path to file with keysets.
+         * @returns {Promise}
+         * @private
+         */
+        _readKeysetsFile: function (filename) {
+            var node = this.node,
+                root = node.getRootDir(),
+                cache = node.getNodeCache(this._target);
+
+            return keysets.read(filename, cache, root);
         }
     })
     .createTech();
