+++ test/mocha.opts
@@ -2,4 +2,5 @@ test/**/*.test.js
 --ui bdd
 --reporter dot
 --require must
+--require test/lib/mocks
 --timeout 5000
