+++ test/techs/bemhtml-18n/bem-bl.test.js
@@ -1,4 +1,5 @@
-var path = require('path'),
+var EOL = require('os').EOL,
+    path = require('path'),
     fs = require('fs'),
     mock = require('mock-fs'),
     serializeJS = require('serialize-javascript'),
@@ -6,12 +7,14 @@ var path = require('path'),
     FileList = require('enb/lib/file-list'),
     dropRequireCache = require('enb/lib/fs/drop-require-cache'),
     Tech = require('../../../techs/bemhtml-i18n'),
-    core = require('../../fixtures/bem-core-v3/common.blocks/i18n/i18n.i18n.js').i18n.i18n,
-    bemhtmlContents;
+    bemhtmlContents,
+    core;
 
-describe('bemxjst bemhtml-i18n v2', function () {
+describe('bemhtml-i18n for bem-bl', function () {
     before(function () {
-        var bemhtmlFilename = './test/fixtures/bem-core/common.blocks/i-bem/i-bem.bemhtml';
+        var coreFilename = './test/fixtures/bem-core/common.blocks/i-bem/__i18n/i-bem__i18n.i18n/core.js',
+            bemhtmlFilename = './test/fixtures/bem-core/common.blocks/i-bem/i-bem.bemhtml';
+        core = fs.readFileSync(path.resolve(coreFilename), { encoding: 'utf-8' });
         bemhtmlContents = fs.readFileSync(path.resolve(bemhtmlFilename), { encoding: 'utf-8' });
     });
 
@@ -19,7 +22,7 @@ describe('bemxjst bemhtml-i18n v2', function () {
         mock.restore();
     });
 
-    it('must throw err if i18n is not found', function () {
+    it('must throw err if i18n core is not found', function () {
         var keysets = {};
 
         return build(keysets)
@@ -29,10 +32,24 @@ describe('bemxjst bemhtml-i18n v2', function () {
             });
     });
 
-    it('must throw err if i18n is not function', function () {
+    it('must throw err if i18n core is not string', function () {
         var keysets = {
-            i18n: {
-                i18n: 'val'
+            all: {
+                '': function () {}
+            }
+        };
+
+        return build(keysets)
+            .fail(function (err) {
+                err.must.a(Error);
+                err.message.must.be('Core of i18n is not found!');
+            });
+    });
+
+    it('must throw err if i18n core is not valid', function () {
+        var keysets = {
+            all: {
+                '': 'hello world'
             }
         };
 
@@ -45,8 +62,8 @@ describe('bemxjst bemhtml-i18n v2', function () {
 
     it('must return value', function () {
         var keysets = {
-            i18n: {
-                i18n: core
+            all: {
+                '': core
             },
             scope: {
                 key: 'val'
@@ -63,11 +80,9 @@ describe('bemxjst bemhtml-i18n v2', function () {
             });
     });
 
-    it('must build fake key if keysets is empty', function () {
+    it('must return empty localization value for empty keysets (only core)', function () {
         var keysets = {
-            i18n: {
-                i18n: core
-            }
+            all: { '': core }
         };
 
         return build(keysets)
@@ -76,59 +91,27 @@ describe('bemxjst bemhtml-i18n v2', function () {
                     bemjson = { block: 'block', scope: 'scope', key: 'key' },
                     html = BEMHTML.apply(bemjson);
 
-                html.must.be('<div class=\"block\">scope:key</div>');
+                html.must.be('<div class=\"block\"></div>');
             });
     });
 
     it('must build key by params', function () {
         var keysets = {
-            i18n: {
-                i18n: core
+            all: {
+                '': core
             },
             scope: {
-                key: function (params) {
-                    return params.join();
-                }
+                key: '<i18n:param>param</i18n:param> value'
             }
         };
 
         return build(keysets)
             .then(function (exports) {
                 var BEMHTML = exports.BEMHTML,
-                    bemjson = { block: 'block', scope: 'scope', key: 'key', params: ['p1', 'p2'] },
+                    bemjson = { block: 'block', scope: 'scope', key: 'key', params: { param: 1 } },
                     html = BEMHTML.apply(bemjson);
 
-                html.must.be('<div class=\"block\">p1,p2</div>');
-            });
-    });
-
-    it('must provide i18n instance to function', function () {
-        var keysets = {
-            i18n: {
-                i18n: core
-            },
-            'scope-1': {
-                key: 'val'
-            },
-            'scope-2': {
-                key: function (params, i18n) {
-                    return i18n(params.scope, params.key);
-                }
-            }
-        };
-
-        return build(keysets)
-            .then(function (exports) {
-                var BEMHTML = exports.BEMHTML,
-                    bemjson = {
-                        block: 'block',
-                        scope: 'scope-1',
-                        key: 'key',
-                        params: { scope: 'scope-1', key: 'key' }
-                    },
-                    html = BEMHTML.apply(bemjson);
-
-                html.must.be('<div class=\"block\">val</div>');
+                html.must.be('<div class=\"block\">1 value</div>');
             });
     });
 
@@ -140,7 +123,7 @@ describe('bemxjst bemhtml-i18n v2', function () {
                 bundle: {
                     'bundle.keysets.lang.js': mock.file({
                         content: serialize({
-                            i18n: { i18n: core },
+                            all: { '': core },
                             scope: { key: 'val' }
                         }),
                         mtime: time
@@ -151,11 +134,13 @@ describe('bemxjst bemhtml-i18n v2', function () {
             var bundle = new MockNode('bundle'),
                 cache = bundle.getNodeCache('bundle.bemhtml.lang.js'),
                 basename = 'bundle.keysets.lang.js',
-                filename = path.resolve('bundle', basename);
+                relPath = path.join('bundle', basename),
+                cacheKey = 'keysets-file-' + relPath,
+                filename = path.resolve(relPath);
 
             dropRequireCache(require, filename);
             require(filename);
-            cache.cacheFileInfo('keysets-file-' + basename, filename);
+            cache.cacheFileInfo(cacheKey, filename);
 
             mock({
                 blocks: {
@@ -165,12 +150,12 @@ describe('bemxjst bemhtml-i18n v2', function () {
                         '    var ctx = this.ctx;',
                         '    return this.i18n(ctx.scope, ctx.key, ctx.params);',
                         '});'
-                    ].join('\n')
+                    ].join(EOL)
                 },
                 bundle: {
                     'bundle.keysets.lang.js': mock.file({
                         content: serialize({
-                            i18n: { i18n: core },
+                            all: { '': core },
                             scope: { key: 'val2' }
                         }),
                         mtime: time
@@ -203,7 +188,7 @@ describe('bemxjst bemhtml-i18n v2', function () {
                 bundle: {
                     'bundle.keysets.lang.js': mock.file({
                         content: serialize({
-                            i18n: { i18n: core },
+                            all: { '': core },
                             scope: { key: 'val' }
                         }),
                         mtime: new Date(1)
@@ -214,11 +199,13 @@ describe('bemxjst bemhtml-i18n v2', function () {
             var bundle = new MockNode('bundle'),
                 cache = bundle.getNodeCache('bundle.bemhtml.lang.js'),
                 basename = 'bundle.keysets.lang.js',
-                filename = path.resolve('bundle', basename);
+                relPath = path.join('bundle', basename),
+                cacheKey = 'keysets-file-' + relPath,
+                filename = path.resolve(relPath);
 
             dropRequireCache(require, filename);
             require(filename);
-            cache.cacheFileInfo('keysets-file-' + basename, filename);
+            cache.cacheFileInfo(cacheKey, filename);
 
             mock({
                 blocks: {
@@ -228,12 +215,12 @@ describe('bemxjst bemhtml-i18n v2', function () {
                         '    var ctx = this.ctx;',
                         '    return this.i18n(ctx.scope, ctx.key, ctx.params);',
                         '});'
-                    ].join('\n')
+                    ].join(EOL)
                 },
                 bundle: {
                     'bundle.keysets.lang.js': mock.file({
                         content: serialize({
-                            i18n: { i18n: core },
+                            all: { '': core },
                             scope: { key: 'val2' }
                         }),
                         mtime: new Date(2)
@@ -272,7 +259,7 @@ function build(keysets) {
                 '    var ctx = this.ctx;',
                 '    return this.i18n(ctx.scope, ctx.key, ctx.params);',
                 '});'
-            ].join('\n')
+            ].join(EOL)
         },
         bundle: {
             'bundle.keysets.lang.js': serialize(keysets)
